feat(socket): add non-blocking socket operations and refactor accept logic
