+++ libzappy_net/Makefile
@@ -32,6 +32,7 @@ HEADERS = $(INC_DIR)/libzappy_net.h \
 SRC_C = $(SRC_DIR)/libzappy_net.c \
         $(SRC_DIR)/zappy_net_core.c \
         $(SRC_DIR)/zappy_net_socket.c \
+        $(SRC_DIR)/zappy_net_socket_ops.c \
         $(SRC_DIR)/zappy_net_server.c \
         $(SRC_DIR)/zappy_net_client.c \
         $(SRC_DIR)/zappy_net_poll.c \
