+++ libzappy_net/src/zappy_net_server.c
@@ -72,26 +72,21 @@ zn_socket_t zn_server_listen(int port)
     return sock;
 }
 
-zn_socket_t zn_accept(zn_socket_t server_sock,
-    struct sockaddr *addr, socklen_t *len)
+static void setup_default_addr_params(struct sockaddr **addr,
+    socklen_t **len, struct sockaddr_in *client_addr, socklen_t *addr_len)
+{
+    if (*addr == NULL || *len == NULL) {
+        *addr = (struct sockaddr *)client_addr;
+        *addr_len = sizeof(*client_addr);
+        *len = addr_len;
+    }
+}
+
+static zn_socket_t create_client_socket(int client_fd,
+    struct sockaddr *addr, socklen_t addr_len)
 {
     zn_socket_t client_sock;
-    int client_fd;
-    socklen_t addr_len;
-    struct sockaddr_in client_addr;
 
-    if (server_sock == NULL || server_sock->fd < 0) {
-        return NULL;
-    }
-    if (addr == NULL || len == NULL) {
-        addr = (struct sockaddr *)&client_addr;
-        addr_len = sizeof(client_addr);
-        len = &addr_len;
-    }
-    client_fd = accept(server_sock->fd, addr, len);
-    if (client_fd < 0) {
-        return NULL;
-    }
     client_sock = zn_socket_create();
     if (client_sock == NULL) {
         close(client_fd);
@@ -100,10 +95,24 @@ zn_socket_t zn_accept(zn_socket_t server_sock,
     client_sock->fd = client_fd;
     client_sock->initialized = 1;
     client_sock->type = 0;
-    if (addr == (struct sockaddr *)&client_addr) {
-        memcpy(&client_sock->addr, &client_addr, sizeof(client_addr));
-    } else {
-        memcpy(&client_sock->addr, addr, *len);
-    }
+    memcpy(&client_sock->addr, addr, addr_len);
     return client_sock;
+}
+
+zn_socket_t zn_accept(zn_socket_t server_sock,
+    struct sockaddr *addr, socklen_t *len)
+{
+    int client_fd;
+    socklen_t addr_len;
+    struct sockaddr_in client_addr;
+
+    if (server_sock == NULL || server_sock->fd < 0) {
+        return NULL;
+    }
+    setup_default_addr_params(&addr, &len, &client_addr, &addr_len);
+    client_fd = accept(server_sock->fd, addr, len);
+    if (client_fd < 0) {
+        return NULL;
+    }
+    return create_client_socket(client_fd, addr, *len);
 }
